Added comments for future references
Program is now fully functional. Users can select either heads or tails, and will have a GUI to select from. Future programs need to deal with error handling, as there are many ways to get past the GUI. Looking to beautify the program a little bit as it's quite ugly, but it functions well.